+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/mobile/diagnostic_message_response.h
@@ -31,37 +31,38 @@
  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_GET_VEHICLE_DATA_RESPONSE_H_
-#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_GET_VEHICLE_DATA_RESPONSE_H_
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_DIAGNOSTIC_MESSAGE_RESPONSE_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_DIAGNOSTIC_MESSAGE_RESPONSE_H_
 
 #include "application_manager/commands/command_response_impl.h"
 #include "utils/macro.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 namespace app_mngr = application_manager;
 
 namespace commands {
 
 /**
- * @brief GetVehicleDataResponse command class
+ * @brief DiagnosticMessageResponse command class
  **/
-class GetVehicleDataResponse : public app_mngr::commands::CommandResponseImpl {
+class DiagnosticMessageResponse
+    : public app_mngr::commands::CommandResponseImpl {
  public:
   /**
-   * @brief GetVehicleDataResponse class constructor
+   * @brief DiagnosticMessageResponse class constructor
    *
    * @param message Incoming SmartObject message
    **/
-  GetVehicleDataResponse(const app_mngr::commands::MessageSharedPtr& message,
-                         app_mngr::ApplicationManager& application_manager,
-                         app_mngr::rpc_service::RPCService& rpc_service,
-                         app_mngr::HMICapabilities& hmi_capabilities,
-                         policy::PolicyHandlerInterface& policy_handler);
+  DiagnosticMessageResponse(const app_mngr::commands::MessageSharedPtr& message,
+                            app_mngr::ApplicationManager& application_manager,
+                            app_mngr::rpc_service::RPCService& rpc_service,
+                            app_mngr::HMICapabilities& hmi_capabilities,
+                            policy::PolicyHandlerInterface& policy_handler);
 
   /**
-   * @brief GetVehicleDataResponse class destructor
+   * @brief DiagnosticMessageResponse class destructor
    **/
-  virtual ~GetVehicleDataResponse();
+  virtual ~DiagnosticMessageResponse();
 
   /**
    * @brief Execute command
@@ -69,10 +70,10 @@ class GetVehicleDataResponse : public app_mngr::commands::CommandResponseImpl {
   virtual void Run();
 
  private:
-  DISALLOW_COPY_AND_ASSIGN(GetVehicleDataResponse);
+  DISALLOW_COPY_AND_ASSIGN(DiagnosticMessageResponse);
 };
 
 }  // namespace commands
 }  // namespace application_manager
 
-#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_GET_VEHICLE_DATA_RESPONSE_H_
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_DIAGNOSTIC_MESSAGE_RESPONSE_H_
