+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/mobile/unsubscribe_vehicle_data_response.h
@@ -31,37 +31,39 @@
  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_GET_DTCS_RESPONSE_H_
-#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_GET_DTCS_RESPONSE_H_
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_UNSUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_UNSUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
 
 #include "application_manager/commands/command_response_impl.h"
 #include "utils/macro.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 namespace app_mngr = application_manager;
 
 namespace commands {
 
 /**
- * @brief GetDTCsResponse command class
+ * @brief UnsubscribeVehicleDataResponse command class
  **/
-class GetDTCsResponse : public app_mngr::commands::CommandResponseImpl {
+class UnsubscribeVehicleDataResponse
+    : public app_mngr::commands::CommandResponseImpl {
  public:
   /**
-   * @brief GetDTCsResponse class constructor
+   * @brief UnsubscribeVehicleDataResponse class constructor
    *
    * @param message Incoming SmartObject message
    **/
-  GetDTCsResponse(const app_mngr::commands::MessageSharedPtr& message,
-                  app_mngr::ApplicationManager& application_manager,
-                  app_mngr::rpc_service::RPCService& rpc_service,
-                  app_mngr::HMICapabilities& hmi_capabilities,
-                  policy::PolicyHandlerInterface& policy_handler);
+  UnsubscribeVehicleDataResponse(
+      const app_mngr::commands::MessageSharedPtr& message,
+      app_mngr::ApplicationManager& application_manager,
+      app_mngr::rpc_service::RPCService& rpc_service,
+      app_mngr::HMICapabilities& hmi_capabilities,
+      policy::PolicyHandlerInterface& policy_handler);
 
   /**
-   * @brief GetDTCsResponse class destructor
+   * @brief UnsubscribeVehicleDataResponse class destructor
    **/
-  virtual ~GetDTCsResponse();
+  virtual ~UnsubscribeVehicleDataResponse();
 
   /**
    * @brief Execute command
@@ -69,10 +71,10 @@ class GetDTCsResponse : public app_mngr::commands::CommandResponseImpl {
   virtual void Run();
 
  private:
-  DISALLOW_COPY_AND_ASSIGN(GetDTCsResponse);
+  DISALLOW_COPY_AND_ASSIGN(UnsubscribeVehicleDataResponse);
 };
 
 }  // namespace commands
 }  // namespace application_manager
 
-#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_GET_DTCS_RESPONSE_H_
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_MOBILE_UNSUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
