+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/vehicle_info_command_factory.h
@@ -68,8 +68,8 @@ private:
   app_mngr::rpc_service::RPCService& rpc_service_;
   app_mngr::HMICapabilities& hmi_capabilities_;
   policy::PolicyHandlerInterface& policy_handler_;
-  std::unique_ptr<VehicleInfoHmiCommandFactory> hmi_command_factory_;
-  std::unique_ptr<VehicleInfoMobileCommandFactory> mob_command_factory_;
+  std::unique_ptr<app_mngr::CommandFactory> hmi_command_factory_;
+  std::unique_ptr<app_mngr::CommandFactory> mob_command_factory_;
 };
 
 }
