+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/vehicle_info_mobile_command_factory.h
@@ -64,6 +64,8 @@ private:
   app_mngr::HMICapabilities& hmi_capabilities_;
   policy::PolicyHandlerInterface& policy_handler_;
 
+  app_mngr::CommandCreator& buildCommandCreator(int32_t function_id, int32_t message_type);
+
   DISALLOW_COPY_AND_ASSIGN(VehicleInfoMobileCommandFactory);
 };
