+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/src/commands/mobile/diagnostic_message_request.cc
@@ -32,14 +32,14 @@
  */
 
 #include <algorithm>
-#include "sdl_rpc_plugin/commands/mobile/diagnostic_message_request.h"
+#include "vehicle_info_plugin/commands/mobile/diagnostic_message_request.h"
 
 #include "application_manager/application_impl.h"
 
 #include "interfaces/HMI_API.h"
 #include "application_manager/message_helper.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 using namespace application_manager;
 
 namespace commands {
