+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/src/commands/mobile/get_dtcs_request.cc
@@ -1,5 +1,6 @@
 /*
- Copyright (c) 2016, Ford Motor Company
+
+ Copyright (c) 2013, Ford Motor Company
  All rights reserved.
 
  Redistribution and use in source and binary forms, with or without
@@ -30,24 +31,22 @@
  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <string>
-#include "sdl_rpc_plugin/commands/mobile/read_did_request.h"
+#include "vehicle_info_plugin/commands/mobile/get_dtcs_request.h"
 
 #include "application_manager/application_impl.h"
-#include "interfaces/MOBILE_API.h"
 #include "interfaces/HMI_API.h"
 #include "application_manager/message_helper.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 using namespace application_manager;
 
 namespace commands {
 
-ReadDIDRequest::ReadDIDRequest(
+GetDTCsRequest::GetDTCsRequest(
     const application_manager::commands::MessageSharedPtr& message,
     ApplicationManager& application_manager,
-    app_mngr::rpc_service::RPCService& rpc_service,
-    app_mngr::HMICapabilities& hmi_capabilities,
+    rpc_service::RPCService& rpc_service,
+    HMICapabilities& hmi_capabilities,
     policy::PolicyHandlerInterface& policy_handler)
     : CommandRequestImpl(message,
                          application_manager,
@@ -55,62 +54,48 @@ ReadDIDRequest::ReadDIDRequest(
                          hmi_capabilities,
                          policy_handler) {}
 
-ReadDIDRequest::~ReadDIDRequest() {}
+GetDTCsRequest::~GetDTCsRequest() {}
 
-void ReadDIDRequest::Run() {
+void GetDTCsRequest::Run() {
   LOG4CXX_AUTO_TRACE(logger_);
 
-  uint32_t app_id =
-      (*message_)[strings::params][strings::connection_key].asUInt();
-
-  ApplicationSharedPtr app = application_manager_.application(app_id);
-  LOG4CXX_INFO(
-      logger_,
-      "Correlation_id :"
-          << (*message_)[strings::params][strings::correlation_id].asUInt());
+  ApplicationSharedPtr app = application_manager_.application(
+      (*message_)[strings::params][strings::connection_key].asUInt());
 
   if (!app) {
-    LOG4CXX_ERROR(logger_, "An application is not registered.");
+    LOG4CXX_ERROR(logger_, "NULL pointer");
     SendResponse(false, mobile_apis::Result::APPLICATION_NOT_REGISTERED);
-    return;
-  }
-
-  if (app->AreCommandLimitsExceeded(
-          static_cast<mobile_apis::FunctionID::eType>(function_id()),
-          application_manager::TLimitSource::CONFIG_FILE)) {
-    LOG4CXX_ERROR(logger_, "ReadDID frequency is too high.");
-    SendResponse(false, mobile_apis::Result::REJECTED);
-    return;
-  }
-
-  if ((*message_)[strings::msg_params][strings::did_location].empty()) {
-    LOG4CXX_ERROR(logger_, "INVALID_DATA");
-    SendResponse(false, mobile_apis::Result::INVALID_DATA);
     return;
   }
 
   smart_objects::SmartObject msg_params =
       smart_objects::SmartObject(smart_objects::SmartType_Map);
-  msg_params[strings::app_id] = app->app_id();
+
   msg_params[strings::ecu_name] =
       (*message_)[strings::msg_params][strings::ecu_name];
-  msg_params[strings::did_location] =
-      (*message_)[strings::msg_params][strings::did_location];
-  StartAwaitForInterface(HmiInterfaces::HMI_INTERFACE_VehicleInfo);
 
-  SendHMIRequest(hmi_apis::FunctionID::VehicleInfo_ReadDID, &msg_params, true);
+  if ((*message_)[strings::msg_params].keyExists(strings::dtc_mask)) {
+    msg_params[strings::dtc_mask] =
+        (*message_)[strings::msg_params][strings::dtc_mask];
+  }
+
+  msg_params[strings::app_id] = app->app_id();
+
+  StartAwaitForInterface(HmiInterfaces::HMI_INTERFACE_VehicleInfo);
+  SendHMIRequest(hmi_apis::FunctionID::VehicleInfo_GetDTCs, &msg_params, true);
 }
 
-void ReadDIDRequest::on_event(const event_engine::Event& event) {
+void GetDTCsRequest::on_event(const event_engine::Event& event) {
   LOG4CXX_AUTO_TRACE(logger_);
   const smart_objects::SmartObject& message = event.smart_object();
 
   switch (event.id()) {
-    case hmi_apis::FunctionID::VehicleInfo_ReadDID: {
+    case hmi_apis::FunctionID::VehicleInfo_GetDTCs: {
       EndAwaitForInterface(HmiInterfaces::HMI_INTERFACE_VehicleInfo);
       hmi_apis::Common_Result::eType result_code =
           static_cast<hmi_apis::Common_Result::eType>(
               message[strings::params][hmi_response::code].asInt());
+
       const bool result = PrepareResultForMobileResponse(
           result_code, HmiInterfaces::HMI_INTERFACE_VehicleInfo);
       std::string response_info;
