+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/src/commands/mobile/read_did_request.cc
@@ -1,6 +1,5 @@
 /*
-
- Copyright (c) 2013, Ford Motor Company
+ Copyright (c) 2016, Ford Motor Company
  All rights reserved.
 
  Redistribution and use in source and binary forms, with or without
@@ -31,22 +30,24 @@
  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "sdl_rpc_plugin/commands/mobile/get_dtcs_request.h"
+#include <string>
+#include "vehicle_info_plugin/commands/mobile/read_did_request.h"
 
 #include "application_manager/application_impl.h"
+#include "interfaces/MOBILE_API.h"
 #include "interfaces/HMI_API.h"
 #include "application_manager/message_helper.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 using namespace application_manager;
 
 namespace commands {
 
-GetDTCsRequest::GetDTCsRequest(
+ReadDIDRequest::ReadDIDRequest(
     const application_manager::commands::MessageSharedPtr& message,
     ApplicationManager& application_manager,
-    rpc_service::RPCService& rpc_service,
-    HMICapabilities& hmi_capabilities,
+    app_mngr::rpc_service::RPCService& rpc_service,
+    app_mngr::HMICapabilities& hmi_capabilities,
     policy::PolicyHandlerInterface& policy_handler)
     : CommandRequestImpl(message,
                          application_manager,
@@ -54,48 +55,62 @@ GetDTCsRequest::GetDTCsRequest(
                          hmi_capabilities,
                          policy_handler) {}
 
-GetDTCsRequest::~GetDTCsRequest() {}
+ReadDIDRequest::~ReadDIDRequest() {}
 
-void GetDTCsRequest::Run() {
+void ReadDIDRequest::Run() {
   LOG4CXX_AUTO_TRACE(logger_);
 
-  ApplicationSharedPtr app = application_manager_.application(
-      (*message_)[strings::params][strings::connection_key].asUInt());
+  uint32_t app_id =
+      (*message_)[strings::params][strings::connection_key].asUInt();
+
+  ApplicationSharedPtr app = application_manager_.application(app_id);
+  LOG4CXX_INFO(
+      logger_,
+      "Correlation_id :"
+          << (*message_)[strings::params][strings::correlation_id].asUInt());
 
   if (!app) {
-    LOG4CXX_ERROR(logger_, "NULL pointer");
+    LOG4CXX_ERROR(logger_, "An application is not registered.");
     SendResponse(false, mobile_apis::Result::APPLICATION_NOT_REGISTERED);
+    return;
+  }
+
+  if (app->AreCommandLimitsExceeded(
+          static_cast<mobile_apis::FunctionID::eType>(function_id()),
+          application_manager::TLimitSource::CONFIG_FILE)) {
+    LOG4CXX_ERROR(logger_, "ReadDID frequency is too high.");
+    SendResponse(false, mobile_apis::Result::REJECTED);
+    return;
+  }
+
+  if ((*message_)[strings::msg_params][strings::did_location].empty()) {
+    LOG4CXX_ERROR(logger_, "INVALID_DATA");
+    SendResponse(false, mobile_apis::Result::INVALID_DATA);
     return;
   }
 
   smart_objects::SmartObject msg_params =
       smart_objects::SmartObject(smart_objects::SmartType_Map);
-
+  msg_params[strings::app_id] = app->app_id();
   msg_params[strings::ecu_name] =
       (*message_)[strings::msg_params][strings::ecu_name];
-
-  if ((*message_)[strings::msg_params].keyExists(strings::dtc_mask)) {
-    msg_params[strings::dtc_mask] =
-        (*message_)[strings::msg_params][strings::dtc_mask];
-  }
-
-  msg_params[strings::app_id] = app->app_id();
-
+  msg_params[strings::did_location] =
+      (*message_)[strings::msg_params][strings::did_location];
   StartAwaitForInterface(HmiInterfaces::HMI_INTERFACE_VehicleInfo);
-  SendHMIRequest(hmi_apis::FunctionID::VehicleInfo_GetDTCs, &msg_params, true);
+
+  SendHMIRequest(hmi_apis::FunctionID::VehicleInfo_ReadDID, &msg_params, true);
 }
 
-void GetDTCsRequest::on_event(const event_engine::Event& event) {
+void ReadDIDRequest::on_event(const event_engine::Event& event) {
   LOG4CXX_AUTO_TRACE(logger_);
   const smart_objects::SmartObject& message = event.smart_object();
 
   switch (event.id()) {
-    case hmi_apis::FunctionID::VehicleInfo_GetDTCs: {
+    case hmi_apis::FunctionID::VehicleInfo_ReadDID: {
       EndAwaitForInterface(HmiInterfaces::HMI_INTERFACE_VehicleInfo);
       hmi_apis::Common_Result::eType result_code =
           static_cast<hmi_apis::Common_Result::eType>(
               message[strings::params][hmi_response::code].asInt());
-
       const bool result = PrepareResultForMobileResponse(
           result_code, HmiInterfaces::HMI_INTERFACE_VehicleInfo);
       std::string response_info;
